--- src/pnbd_expectation.h
@@ -1,10 +1,0 @@
-#ifndef PNBD_EXPECTATION_HPP
-#define PNBD_EXPECTATION_HPP
-
-arma::vec pnbd_expectation(const double r,
-                              const double s,
-                              const arma::vec& vAlpha_i,
-                              const arma::vec& vBeta_i,
-                              const arma::vec& vT_i);
-
-#endif
